支持local类型，直接播放本地视频；支持 MuseTalk 对接

+++ utils/common.py
@@ -96,13 +96,14 @@ class Common:
             now_fmt = beijing_now.strftime(fmt)
             return now_fmt
     
-    def move_and_rename(self, src_file_path, target_dir, new_filename=None):
+    def move_and_rename(self, src_file_path, target_dir, new_filename=None, max_attempts=3):
         """
         移动文件到指定目录，可选地重命名文件。
         
         :param src_file_path: 源文件的完整路径
         :param target_dir: 目标目录的路径
         :param new_filename: 可选的新文件名（不包括目录路径）
+        :param max_attempts: 最大重试次数（默认为3）
         """
 
         import os
@@ -115,11 +116,32 @@ class Common:
         # 如果提供了新文件名，则使用它；否则，保持文件的原始名称
         filename = new_filename if new_filename else os.path.basename(src_file_path)
         target_file_path = os.path.join(target_dir, filename)
+        logging.debug(f"target_dir={target_dir}, target_file_path={target_file_path}")
 
-        # 移动并重命名文件（如有必要）
-        shutil.move(src_file_path, target_file_path)
+        # 尝试移动文件，如果文件被其他程序占用，则进行重试
+        attempts = 0
+        retry_delay = 0.5
 
-        logging.info(f"文件:{src_file_path} 已移动到:{target_file_path}")
+        while attempts < max_attempts:
+            try:
+                # 如果源文件不存在，则直接返回False
+                if not os.path.exists(src_file_path):
+                    logging.error(f"文件移动失败: 源文件{src_file_path}不存在")
+                    return False
+                
+                shutil.move(src_file_path, target_file_path)
+                logging.info(f"文件:{src_file_path} 已移动到:{target_file_path}")
+                return True
+            except Exception as e:
+                logging.error(f"文件移动失败: {e}")
+                logging.info("重试中...")
+                attempts += 1
+                time.sleep(retry_delay)
+
+        logging.warning(f"达到最大重试次数({max_attempts})，无法移动文件:", src_file_path)
+        return False
+
+        
     
     def get_filename_with_ext(self, filepath):
         """
